requirements.txt
aiohttp>=3.8.1
async_timeout>=4.0.2
beautifulsoup4>=4.11.1
Pillow>=9.2.0
pilmoji>=2.0.1
Pyrogram>=2.0.26
PyYAML>=6.0
requests>=2.26.0
